add: Pattern Code Helpers
