+++ CMakeLists.txt
@@ -12,6 +12,7 @@ add_executable(${PROJECT_NAME}
     src/buffer/buffer.c
     src/error/error.c
     src/pattern/code.c
+    src/pattern/decode.c
     src/pattern/hash.c
     src/pattern/matcher.c
     src/pattern/parser.c
