+++ src/pattern/parser.c
@@ -11,71 +11,6 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-static struct ptrn create_edge(ptr target_offset, byte literal, byte other)
-{
-    return (struct ptrn){
-        .type          = EDGE,
-        .target_offset = target_offset,
-        .literal       = literal,
-        .other         = other};
-}
-
-static struct ptrn create_vertex(ptr edges)
-{
-    return (struct ptrn){.type = VERTEX, .edges = edges};
-}
-
-static struct ptrn create_marker(struct str name, bool visible)
-{
-    return (struct ptrn){.type = MARKER, .name = name, .visible = visible};
-}
-
-static ptr size(struct ptrnctx ptrns)
-{
-    ASSERT(ptrns.end >= ptrns.bgn, "Patterns do not have a positive size!");
-    return ptrns.end - ptrns.bgn;
-}
-
-static ptr capacity(struct ptrnctx ptrns)
-{
-    ASSERT(
-        ptrns.lst >= ptrnctx.bgn,
-        "Patterns do not have a positive capacity!");
-    return ptrns.lst - ptrns.bgn;
-}
-
-static ptr space(struct ptrns ptrns)
-{
-    ASSERT(ptrns.lst >= ptrns.end, "Patterns do not have a positive space!");
-    return ptrns.lst - ptrns.end;
-}
-
-static struct ptrnctx grow(struct ptrnctx ptrns)
-{
-    ptr cap  = capacity(ptrns);
-    ptr sze  = size(ptrns);
-    ptr grow = max(8, cap / 2);
-    ptr nwc  = grow + cap;
-
-    ASSERT(nwc > 0, "New patterns capacity is not positive!");
-    struct ptrn* mem = realloc(ptrns.bgn, nwc * sizeof(struct ptrn));
-    CHECK(mem != null, "Could not allocate patterns!");
-
-    ptrns.bgn = mem;
-    ptrns.end = mem + sze;
-    ptrns.lst = mem + nwc;
-    return ptrns;
-}
-
-static struct ptrnctx put(struct ptrnctx ptrns, struct ptrn ptrn)
-{
-    if (space(ptrns) < 1) {
-        ptrns = grow(ptrns);
-    }
-    *ptrns.end++ = ptrn;
-    return ptrns;
-}
-
 /* Copy result together with the changed patterns. */
 struct copy_res {
     struct ptrnctx ptrns;
